minor adjustment

+++ hpylib/util/remote_data.py
@@ -28,16 +28,21 @@ def info_host(hostname):
 def sync_from_remote(hostname, filename):
 
     # filename is the file name including the absolute path on remote machine
+
     home_path = os.getenv('HOME')+'/'
     user, url = info_host(hostname)
     dat_pastr = '/data_'+str.lower(hostname)+'/'
     sub_index = filename.find(dat_pastr)
 
     filename_local = home_path+'data_local/'+filename[sub_index+len(dat_pastr):]
+
+    if os.path.isfile(filename_local):
+        return 0
+
     path_index = filename_local.rfind('/')
     path_local = filename_local[0:path_index]
 
     if not os.path.exists(path_local):
         os.makedirs(path_local)
 
-    os.system("scp "+user+"@"+url+":"+filename+" "+filename_local)
+    return os.system("scp "+user+"@"+url+":"+filename+" "+filename_local)
